[i2c: reflect number of bytes transfered]
Update the i2c driver api, in particular the return values of the i2c
read/write/xfer functions, to pass the number of bytes transfered
through to the caller.

This is not used by most of the higher-level drivers, except the raw i2c
device driver, which needs it to correctly communicate to user-space.

+++ include/driver/i2c.h
@@ -108,9 +108,9 @@ typedef struct i2c_t{
 i2c_t *i2c_create(i2c_ops_t *ops, i2c_cfg_t *cfg, void *hw);
 void i2c_destroy(i2c_t *i2c);
 
-int i2c_read(i2c_t *i2c, i2c_mode_t mode, uint8_t slave, void *buf, size_t n);
-int i2c_write(i2c_t *i2c, i2c_mode_t mode, uint8_t slave, void *buf, size_t n);
-int i2c_xfer(i2c_t *i2c, i2c_mode_t mode, i2c_cmd_t cmd, uint8_t slave, blob_t *bufs, size_t n);
+size_t i2c_read(i2c_t *i2c, i2c_mode_t mode, uint8_t slave, void *buf, size_t n);
+size_t i2c_write(i2c_t *i2c, i2c_mode_t mode, uint8_t slave, void *buf, size_t n);
+size_t i2c_xfer(i2c_t *i2c, i2c_mode_t mode, i2c_cmd_t cmd, uint8_t slave, blob_t *bufs, size_t n);
 
 
 #endif // DRIVER_I2C_H
